Fix training pipeline: data paths, OOM, dtype, and robustness
- Use DATA_DIR config resolution for variant_bank_path (fixes Docker mount)
- Add file_read_repro to generate-variants Makefile target
- Reduce max_batch_tokens 65K→16K and chunk cross-entropy loss to fit 128 GB
- Cast encoder parameters to model dtype in from_pretrained (bf16 fix)
- Probe bitsandbytes CUDA kernels before committing to 8-bit optimizer
- Add pool-restart recovery for extract-structural worker crashes
- Add CLAUDECODE env check to generate_prompt_variants.py

+++ tests/unit/models/test_encoder.py
@@ -251,7 +251,8 @@ class TestBgKITEncoderFactory:
 
         encoder = BgKITEncoder.from_pretrained(backbone, hidden_dim=64)
 
-        x = torch.randn(2, 10, 64)
+        # from_pretrained casts to bfloat16; input must match
+        x = torch.randn(2, 10, 64, dtype=torch.bfloat16)
         out = encoder(x, survivor_mask=None)
         assert isinstance(out, CompressionOutput)
         assert out.survivor_embeddings.shape == (2, 10, 64)
